Changes the Windows workaround created for this test case.
This is a strange behavior that probably occurs under some combinations of Windows versions and JDK versions.

This integration changes the implementation of this test so that it fails (as the test expects) early if it is impossible to make the directory being tested read only.

I have two different Windows 2008 R2 systems the original version passes on one, and fails on the other. Go figure.