Adds camel endpoint support for MPQ
There are 2 tests that FAIL in this commit:
    TestIti51.testIti51
    TestIti51.testIti51FailureAudit

These are being checked in as is to hopefully get a code review performed. In the mean time, this feature is broken without proper audit capability so I will leave these tests as is for now until I can get better guidance from the IPF team.